Add basic example input files
There's no documentation beyond the source code itself, so we should
have some example input files so people can actually run the program.

This commit adds some input files that show a couple of the molecules
from the original paper, along with displaying the basis option. It's
not really enough, because we should show the rest of the options too.
But it is a start.